Create README.md
I'm a woman in a man's area, which is fine because I worked in construction for 20 years.  Because I have seen so few women in this field, it is a challenge I'm looking forward to.  Visit my internet channels on any of the main social media platforms under seo services, seo specialist or Jill Pierce.